+++ src/lightning_app/components/training.py
@@ -121,7 +121,7 @@ class LightningTrainingComponent(LightningFlow):
         script_runner: Type[TracerPythonScript] = PyTorchLightningScriptRunner,
         **script_runner_kwargs,
     ):
-        """This component enables to perform distributed multi-node multi-devices training.
+        """This component enables performing distributed multi-node multi-device training.
 
         Example::
 
@@ -142,7 +142,7 @@ class LightningTrainingComponent(LightningFlow):
             script_args: The arguments to be pass to the script.
             num_nodes: Number of nodes.
             cloud_compute: The cloud compute object used in the cloud.
-            sanity_serving: Whether to validate the model correctly implements
+            sanity_serving: Whether to validate that the model correctly implements
                 the ServableModule API
         """
         super().__init__()